# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DM Interactive Table API
  version: 2.0.0-alpha.23
  description: API HTTP para sesiones, campañas, escenas, mapas, combate, dados e IA.
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
  - dmCookie: []
  - tableCookie: []
paths:
...
  /table-access:
    get:
      summary: Consultar el acceso temporal de una campaña
      operationId: getTableAccess
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Estado del acceso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableAccessStatus"
    post:
      summary: Generar un código temporal de mesa
      operationId: createTableAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [campaignId]
              properties:
                campaignId:
                  type: string
                sessionId:
                  type: string
                playerEnabled:
                  type: boolean
                  default: true
                displayEnabled:
                  type: boolean
                  default: true
                ttlMinutes:
                  type: integer
                  minimum: 15
                  maximum: 1440
      responses:
        "201":
          description: Código generado y código anterior revocado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableAccessGrant"
    delete:
      summary: Revocar el acceso temporal de una campaña
      operationId: revokeTableAccess
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Acceso revocado
...
  /campaigns/{campaignId}/sessions:
    get:
      summary: Listar sesiones de campaña
      operationId: listCampaignSessions
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Lista de sesiones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSession"
    post:
      summary: Crear sesión
      operationId: createCampaignSession
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameSessionRequest"
      responses:
        "201":
          description: Sesión creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
  /campaigns/{campaignId}/workflow:
    get:
      summary: Consultar lobby, sesiones y presencia de una campaña
      operationId: getSessionWorkflow
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Estado actual del workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWorkflowState"
  /campaigns/{campaignId}/workflow/sessions/{sessionId}/start:
    post:
      summary: Iniciar formalmente una sesión
      operationId: startGameSession
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Sesión en vivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLifecycleResponse"
        "409":
          description: Otra sesión ya está en vivo
  /campaigns/{campaignId}/workflow/sessions/{sessionId}/end:
    post:
      summary: Finalizar una sesión y guardar sus resúmenes
      operationId: endGameSession
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                summaryPublic:
                  type: string
                  maxLength: 5000
                summaryPrivate:
                  type: string
                  maxLength: 5000
      responses:
        "200":
          description: Sesión finalizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLifecycleResponse"
  /campaigns/{campaignId}/workflow/sessions/{sessionId}/reopen:
    post:
      summary: Devolver una sesión finalizada a preparación
      operationId: reopenGameSession
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Sesión reabierta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLifecycleResponse"
  /campaigns/{campaignId}/workflow/character-sheet:
    get:
      summary: Consultar la hoja del jugador autenticado
      operationId: getMyCharacterSheet
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Hoja persistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerCharacterSheet"
    patch:
      summary: Actualizar la hoja del jugador autenticado
      operationId: updateMyCharacterSheet
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerCharacterSheet"
      responses:
        "200":
          description: Hoja actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerCharacterSheet"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dmCookie:
      type: apiKey
      in: cookie
      name: dit_dm_session
    tableCookie:
      type: apiKey
      in: cookie
      name: dit_table_session
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    TableAccessGrant:
      type: object
      required:
        [code, campaignId, playerEnabled, displayEnabled, expiresAt]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9]{6}$"
        campaignId:
          type: string
        sessionId:
          type: string
        playerEnabled:
          type: boolean
        displayEnabled:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    TableAccessStatus:
      type: object
      required: [active, campaignId]
      properties:
        active:
          type: boolean
        campaignId:
          type: string
        sessionId:
          type: string
        playerEnabled:
          type: boolean
        displayEnabled:
          type: boolean
        expiresAt:
          type: string
          format: date-time
...
    Campaign:
      type: object
      required: [id, name, ruleset, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ruleset:
          type: string
          example: dnd5e
        status:
          type: string
          enum: [draft, active, archived]
...
    GameSession:
      type: object
      required: [id, campaignId, title, phase, summaryPublic, summaryPrivate]
      properties:
        id:
          type: string
        campaignId:
          type: string
        title:
          type: string
        phase:
          type: string
          enum: [preparation, live, ended]
        scheduledAt:
          type: string
        summaryPublic:
          type: string
        summaryPrivate:
          type: string
    CreateGameSessionRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        scheduledAt:
          type: string
    SessionParticipant:
      type: object
      required: [id, displayName, role, connectedAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [dm, player, display]
        connectedAt:
          type: string
          format: date-time
    SessionWorkflowState:
      type: object
      required: [campaign, sessions, participantsBySession]
      properties:
        campaign:
          $ref: "#/components/schemas/Campaign"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/GameSession"
        participantsBySession:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/SessionParticipant"
    SessionLifecycleResponse:
      type: object
      required: [session]
      properties:
        session:
          $ref: "#/components/schemas/GameSession"
    PlayerCharacterSheet:
      type: object
      required:
        - id
        - campaignId
        - playerKey
        - name
        - ancestry
        - className
        - level
        - maxHp
        - currentHp
        - temporaryHp
        - armorClass
        - notes
        - resources
        - updatedAt
      properties:
        id:
          type: string
        campaignId:
          type: string
        playerKey:
          type: string
        name:
          type: string
        ancestry:
          type: string
        className:
          type: string
        level:
          type: integer
        maxHp:
          type: integer
        currentHp:
          type: integer
        temporaryHp:
          type: integer
        armorClass:
          type: integer
        notes:
          type: string
        resources:
          type: object
          additionalProperties:
            type: integer
        updatedAt:
          type: string
          format: date-time
    UpdatePlayerCharacterSheet:
      type: object
      properties:
        name:
          type: string
        ancestry:
          type: string
        className:
          type: string
        level:
          type: integer
        maxHp:
          type: integer
        currentHp:
          type: integer
        temporaryHp:
          type: integer
        armorClass:
          type: integer
        notes:
          type: string
        resources:
          type: object
          additionalProperties:
            type: integer
...